Move code next to where it’s used
