+++ commands/suggest_stacks.go
@@ -1,10 +1,51 @@
 package commands
 
 import (
+	"html/template"
+
 	"github.com/spf13/cobra"
 
 	"github.com/buildpack/pack/logging"
 )
+
+type suggestedStack struct {
+	ID          string
+	Description string
+	Maintainer  string
+	BuildImage  string
+	RunImage    string
+}
+
+var suggestedStacks = []suggestedStack{
+	{
+		ID:          "heroku-18",
+		Description: "The official Heroku stack based on Ubuntu 18.04",
+		Maintainer:  "Heroku",
+		BuildImage:  "heroku/pack:18-build",
+		RunImage:    "heroku/pack:18",
+	},
+	{
+		ID:          "io.buildpacks.stacks.bionic",
+		Description: "A minimal Cloud Foundry stack based on Ubuntu 18.04",
+		Maintainer:  "Cloud Foundry",
+		BuildImage:  "cloudfoundry/build:base-cnb",
+		RunImage:    "cloudfoundry/run:base-cnb",
+	},
+	{
+		ID:          "org.cloudfoundry.stacks.cflinuxfs3",
+		Description: "A large Cloud Foundry stack based on Ubuntu 18.04",
+		Maintainer:  "Cloud Foundry",
+		BuildImage:  "cloudfoundry/build:full-cnb",
+		RunImage:    "cloudfoundry/run:full-cnb",
+	},
+	{
+		ID:          "org.cloudfoundry.stacks.tiny",
+		Description: "A tiny Cloud Foundry stack based on Ubuntu 18.04, similar to distroless",
+		Maintainer:  "Cloud Foundry",
+		BuildImage:  "cloudfoundry/build:tiny-cnb",
+		RunImage:    "cloudfoundry/run:tiny-cnb",
+	},
+}
 
 func SuggestStacks(logger logging.Logger) *cobra.Command {
 	cmd := &cobra.Command{
@@ -19,4 +60,19 @@ func SuggestStacks(logger logging.Logger) *cobra.Command {
 
 	AddHelpFlag(cmd, "suggest-stacks")
 	return cmd
+}
+
+func suggestStacks(log logging.Logger) {
+	tmpl := template.Must(template.New("").Parse(`
+Stacks maintained by the community:
+{{- range . }}
+
+    Stack ID: {{ .ID }}
+    Description: {{ .Description }}
+    Maintainer: {{ .Maintainer }}
+    Build Image: {{ .BuildImage }}
+    Run Image: {{ .RunImage }}
+{{- end }}
+`))
+	tmpl.Execute(log.Writer(), suggestedStacks)
 }
